requirements.txt
pyyaml
pydantic
black
gitpython
python-dotenv
requests
zstandard
fastapi
netaddr
uvicorn
